Add log_colorize configuration option
This option allows us disable the colorize of texts in the generated
logs. This is useful when we are using the library on our k8s cluster
and we want to get clearer logs in kibana.

By default it will have a value of true to maintain the same behaviour
when the new option doesn’t be sent.